refactor utility methods
