+++ faq-site/ReadMe.md
@@ -23,7 +23,6 @@ npm run update-and-build  # gets new KB data and builds site
 
 ## // TODO
 
-- [ ] fix focus styles
 - [ ] #26 update timestamp from KB modified time, not build time #26
 - [ ] #37 add toggle for TOC on mobile
 - [ ] add deep links to filter and each fragment URI
@@ -35,6 +34,7 @@ npm run update-and-build  # gets new KB data and builds site
 - [ ] fix [tel:211]
 - [ ] social meta tags
 - [ ] determine multiple env for config (workaround comment out)
+- [x] fix focus styles
 - [x] #39 add back to top button
 - [x] #40 guarantee unique IDs for subcategories
 - [x] handle no js
