--- faq-site/_data/faqs.js
@@ -1,9 +1,0 @@
-const fs = require("fs");
-const jsoncParser = require("jsonc-parser")
-
-let fullPath = `${__dirname}/faqs.jsonc`;
-let contents = fs.readFileSync(fullPath, "utf8")
-let object = jsoncParser.parse(contents)
-
-// jsonc adapter
-module.exports = object
